// client/dist/widgets/arcgis/arcgis-map/src/runtime/components/multisourcemap-context.ts
import { React, type IMThemeVariables } from 'jimu-core'

interface MultiSourceMapContextType {
  mapWidgetId: string
  mapWidgetHeight: number
  isShowMapSwitchBtn: boolean
  isFullScreen: boolean
  dataSourceIds: string[]
  activeDataSourceId: string
  switchMap: () => void
  fullScreenMap: () => void
  initialMapState: any
  mobilePanelContainer: HTMLDivElement
  onMobilePanelContentChange: (MobilePanelContent: JSX.Element) => void
  theme: IMThemeVariables
}

export const MultiSourceMapContext = React.createContext<MultiSourceMapContextType>({
  mapWidgetId: null,
  mapWidgetHeight: null,
  isShowMapSwitchBtn: false,
  isFullScreen: false,
  dataSourceIds: [],
  activeDataSourceId: null,
  switchMap: () => {},
  fullScreenMap: () => {},
  initialMapState: null,
  mobilePanelContainer: null,
  onMobilePanelContentChange: (MobilePanelContent: JSX.Element) => {},
  theme: null
})